Moved server paths to router.
Also moved all the back files to a "back" folder.
