+++ package.json
@@ -6,17 +6,18 @@
   "dependencies": {
     "cors": "^2.8.5",
     "express": "^4.17.1",
+    "mongoose": "^6.0.12",
+    "morgan": "^1.10.0",
     "webpack": "^5.62.1",
-    "webpack-cli": "^4.9.1",
-    "morgan": "^1.10.0"
+    "webpack-cli": "^4.9.1"
   },
   "devDependencies": {
     "nodemon": "^2.0.14"
   },
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
-    "start": "node ./src/server.js",
-    "dev": "nodemon ./src/server.js",
+    "start": "node ./src/back/server.js",
+    "dev": "nodemon ./src/back/server.js",
     "build": "webpack"
   },
   "repository": {
