+++ src/back/router.js
@@ -1,40 +1,25 @@
 const express = require('express');
 const data = require('./data');
-const morgan = require('morgan');
-const cors = require('cors');
 const { getMatchingPerson, checkTakenName } = require('./helpers')
-const app = express();
-const port = process.env.PORT || 3001;
+const router = express.Router();
 
-app.use(cors());
-app.use(express.json());
-app.listen(port, () => {
-    console.log(`Listening on port ${port}...`);
+
+router.use('/', express.static(`${__dirname}/../front`));
+router.use('/', express.static(`${__dirname}/../../dist`));
+router.get('/', (req, res) => {
+    res.sendFile(`${__dirname}/../front/index.html`);
 })
 
-morgan.token("data", (req) => {
-    return JSON.stringify(req.body)
-});
-app.use(morgan(':method :url :status :response-time ms - :res[content-length] :data', {
-    skip: function (req, res) { return req.method !== "POST" }
-}));
-
-app.use('/', express.static(`${__dirname}/front`));
-app.use('/', express.static(`${__dirname}/../dist`));
-app.get('/', (req, res) => {
-    res.sendFile(`${__dirname}/front/index.html`);
-})
-
-app.get('/api/persons', (req, res) => {
+router.get('/api/persons', (req, res) => {
     res.send(data);
 })
 
-app.get('/info', (req, res) => {
+router.get('/info', (req, res) => {
     const phonebookEntries = data.length;
     res.send(`Phonebook has info for ${phonebookEntries} people. \n ${Date()}`);
 })
 
-app.get('/api/persons/:id', (req, res) => {
+router.get('/api/persons/:id', (req, res) => {
     const personId = Number(req.params.id);
     const matchingPerson = (getMatchingPerson(personId, data));
     matchingPerson
@@ -42,7 +27,7 @@ app.get('/api/persons/:id', (req, res) => {
     : res.status(400).send({error: "Can't find person"});
 })
 
-app.delete('/api/persons/:id', (req, res) => {
+router.delete('/api/persons/:id', (req, res) => {
     const personId = Number(req.params.id);
     const matchingPerson = (getMatchingPerson(personId, data));
     if(matchingPerson){
@@ -54,7 +39,7 @@ app.delete('/api/persons/:id', (req, res) => {
     }
 })
 
-app.post('/api/persons', (req, res) => {
+router.post('/api/persons', (req, res) => {
     const generatedId = Math.floor(Math.random()*100000);
     const { name, number } = req.body;
     if(!name || !number){
@@ -69,3 +54,5 @@ app.post('/api/persons', (req, res) => {
     data.push(addPerson);
     res.send(data);
 })
+
+module.exports = router;
